Refactors handlers and their tests
Separated most of handlers to own files inside handlers folder.
Refactored tests for new set up.
Fixed imports in server.py.
