+++ metadata_backend/api/handlers/__init__.py
@@ -1,0 +1,1 @@
+"""API handlers."""
